% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{fun_get_map}
\alias{fun_get_map}
\title{Get country maps}
\usage{
fun_get_map(cntry = "NZ", kp = 0.03, min_area = 0.001, pth = "")
}
\arguments{
\item{cntry}{iso country code, use raster::getData('ISO3') to see codes}

\item{kp}{proportion of original points to keep}

\item{min_area}{min area (km^2) of polygons to keep}

\item{pth}{file path where downloaded map should be saved}
}
\value{
a SpatialPolygon
}
\description{
Import maps of countries from the \href{https://gadm.org}{gadm} database and compress to a chosen file size. File size is reduced by a) reducing the number of points in the polygon(s) and b) removing islands smaller than a chosen size.
}
\examples{
\dontrun{
nz_small <- fun_get_map(cntry = "NZ", kp = 0.03, min_area = 0.001)
}
}
